+++ src/main/php/lang/ast/emit/PHP80.class.php
@@ -1,6 +1,7 @@
 <?php namespace lang\ast\emit;
 
 use lang\ast\Node;
+use lang\ast\types\{IsUnion, IsFunction, IsArray, IsMap};
 
 /**
  * PHP 8.0 syntax
@@ -10,7 +11,33 @@ use lang\ast\Node;
 class PHP80 extends PHP {
   use RewriteBlockLambdaExpressions;
 
-  protected $unsupported= [];
+  /**
+   * Rewrite types:
+   *
+   * - No type or function types - emit unchecked
+   * - Array and map types - emit `array` type hint
+   * - Union types - Apply for all components
+   * - Otherwise, simply use type literal as-is
+   */
+  protected function literal($type) {
+    if (null === $type || $type instanceof IsFunction) {
+      return '';
+    } else if ($type instanceof IsArray || $type instanceof IsMap) {
+      return 'array';
+    } else if ($type instanceof IsUnion) {
+      $literal= '';
+      foreach ($type->components as $component) {
+        $literal.= '|'.$this->paramType($component);
+      }
+      return substr($literal, 1);
+    } else {
+      return $type->literal();
+    }
+  }
+
+  protected function paramType($type) { return $this->literal($type); }
+
+  protected function returnType($type) { return $this->literal($type); }
 
   protected function emitArguments($result, $arguments) {
     $s= sizeof($arguments) - 1;