+++ src/test/php/lang/ast/unittest/emit/UnionTypesTest.class.php
@@ -1,12 +1,13 @@
 <?php namespace lang\ast\unittest\emit;
 
-use lang\{Primitive, TypeUnion};
-use unittest\{Assert, Test};
+use lang\{Primitive, Type, TypeUnion};
+use unittest\actions\RuntimeVersion;
+use unittest\{Assert, Test, Action};
 
 /**
- * Function types
+ * Union types
  *
- * @see  https://wiki.php.net/rfc/union_types (Declined)
+ * @see  https://wiki.php.net/rfc/union_types_v2
  */
 class UnionTypesTest extends EmittingTest {
 
@@ -43,6 +44,30 @@ class UnionTypesTest extends EmittingTest {
     Assert::equals(
       new TypeUnion([Primitive::$INT, Primitive::$STRING]),
       $t->getMethod('test')->getReturnType()
+    );
+  }
+
+  #[Test, Action(eval: 'new RuntimeVersion(">=8.0.0-dev")')]
+  public function parameter_type_restriction_with_php8() {
+    $t= $this->type('class <T> {
+      public function test(int|string|array<string> $arg) { }
+    }');
+
+    Assert::equals(
+      new TypeUnion([Primitive::$INT, Primitive::$STRING, Type::$ARRAY]),
+      $t->getMethod('test')->getParameter(0)->getTypeRestriction()
+    );
+  }
+
+  #[Test, Action(eval: 'new RuntimeVersion(">=8.0.0-dev")')]
+  public function return_type_restriction_with_php8() {
+    $t= $this->type('class <T> {
+      public function test(): int|string|array<string> { }
+    }');
+
+    Assert::equals(
+      new TypeUnion([Primitive::$INT, Primitive::$STRING, Type::$ARRAY]),
+      $t->getMethod('test')->getReturnTypeRestriction()
     );
   }
 }